An use float type

+++ src/renderer/DLR_Visualization/visualize.jl
@@ -36,50 +36,50 @@ function visualizeObject(obj::Composition.Object3D, id::Ptr{Nothing}, simVis::Si
     if shapeKind == Modia3D.SphereKind
         sphere::Modia3D.Sphere = obj.shape
         visualizeShape(simVis, obj.r_abs, obj.R_abs, SimVisSphere, id, obj.visualMaterial,
-                       sphere.diameter, sphere.diameter, sphere.diameter)
+                       Float64(sphere.diameter), Float64(sphere.diameter), Float64(sphere.diameter) )
 
     elseif shapeKind == Modia3D.EllipsoidKind
         ellipsoid::Modia3D.Ellipsoid = obj.shape
         visualizeShape(simVis, obj.r_abs, obj.R_abs, SimVisSphere, id, obj.visualMaterial,
-                       ellipsoid.lengthX, ellipsoid.lengthY, ellipsoid.lengthZ)
+        Float64(ellipsoid.lengthX), Float64(ellipsoid.lengthY), Float64(ellipsoid.lengthZ) )
 
     elseif shapeKind == Modia3D.BoxKind
         box::Modia3D.Box = obj.shape
         visualizeShape(simVis, obj.r_abs, obj.R_abs, SimVisBox, id, obj.visualMaterial,
-                       box.lengthX, box.lengthY, box.lengthZ)
+        Float64(box.lengthX), Float64(box.lengthY), Float64(box.lengthZ) )
 
     elseif shapeKind == Modia3D.CylinderKind
         cylinder::Modia3D.Cylinder = obj.shape
         if cylinder.innerDiameter == 0.0
             visualizeShape(simVis, obj.r_abs, Shapes.rotateAxis2z(cylinder.axis, obj.R_abs), SimVisCylinder, id, obj.visualMaterial,
-                           cylinder.diameter, cylinder.diameter, cylinder.length)
+            Float64(cylinder.diameter), Float64(cylinder.diameter), Float64(cylinder.length) )
         else
             visualizeShape(simVis, obj.r_abs, Shapes.rotateAxis2z(cylinder.axis, obj.R_abs), SimVisPipe, id, obj.visualMaterial,
-                           cylinder.diameter, cylinder.diameter, cylinder.length; extras=@MVector[0.0, cylinder.innerDiameter/cylinder.diameter, 1.0])
+            Float64(cylinder.diameter), Float64(cylinder.diameter), Float64(cylinder.length); extras=@MVector[0.0, Float64(cylinder.innerDiameter/cylinder.diameter), 1.0])
         end
 
     elseif shapeKind == Modia3D.ConeKind
         cone::Modia3D.Cone = obj.shape
         if cone.axis == 1
-            dr = @MVector[cone.length/4, 0.0, 0.0]
+            dr = @MVector[Float64(cone.length/4), 0.0, 0.0]
         elseif cone.axis == 2
-            dr = @MVector[0.0, cone.length/4, 0.0]
+            dr = @MVector[0.0, Float64(cone.length/4), 0.0]
         else
-            dr = @MVector[0.0, 0.0, cone.length/4]
+            dr = @MVector[0.0, 0.0, Float64(cone.length/4)]
         end
         r_abs = obj.r_abs + obj.R_abs' * dr
         visualizeShape(simVis, r_abs, Shapes.rotateAxis2z(cone.axis, obj.R_abs), SimVisCone, id, obj.visualMaterial,
-                       cone.diameter, cone.diameter, cone.length; extras=@MVector[cone.topDiameter/cone.diameter, 0.0, 0.0])
+        Float64(cone.diameter), Float64(cone.diameter), Float64(cone.length); extras=@MVector[Float64(cone.topDiameter/cone.diameter), 0.0, 0.0])
 
     elseif shapeKind == Modia3D.CapsuleKind
         capsule::Modia3D.Capsule = obj.shape
         visualizeShape(simVis, obj.r_abs, Shapes.rotateAxis2z(capsule.axis, obj.R_abs), SimVisCapsule, id, obj.visualMaterial,
-                       capsule.diameter, capsule.diameter, capsule.length)
+        Float64(capsule.diameter), Float64(capsule.diameter), Float64(capsule.length) )
 
     elseif shapeKind == Modia3D.BeamKind
         beam::Modia3D.Beam = obj.shape
         visualizeShape(simVis, obj.r_abs, Shapes.rotateAxis2x(beam.axis, obj.R_abs), SimVisBeam, id, obj.visualMaterial,
-                       beam.length, beam.width, beam.thickness)
+        Float64(beam.length), Float64(beam.width), Float64(beam.thickness) )
 
     elseif shapeKind == Modia3D.CoordinateSystemKind
         coordinateSystem::Modia3D.CoordinateSystem = obj.shape
